+++ src/QueryCat.Backend/Commands/Select/SelectPlanner.QueryContext.cs
@@ -4,6 +4,7 @@ using QueryCat.Backend.Ast.Nodes.Select;
 using QueryCat.Backend.Commands.Select.Visitors;
 using QueryCat.Backend.Functions;
 using QueryCat.Backend.Relational;
+using QueryCat.Backend.Storage;
 using QueryCat.Backend.Types;
 
 namespace QueryCat.Backend.Commands.Select;
@@ -11,13 +12,13 @@ namespace QueryCat.Backend.Commands.Select;
 internal sealed partial class SelectPlanner
 {
     private void QueryContext_FillQueryContextConditions(
-        SelectQuerySpecificationNode querySpecificationNode,
-        SelectCommandContext commandContext)
+        SelectCommandContext context,
+        SelectQuerySpecificationNode querySpecificationNode)
     {
         // Fill conditions.
-        foreach (var context in commandContext.InputQueryContextList)
+        foreach (var inputContext in context.InputQueryContextList)
         {
-            QueryContext_FillQueryContextConditions(querySpecificationNode, context, commandContext);
+            QueryContext_FillQueryContextConditions(querySpecificationNode, inputContext, context);
         }
 
         // Fill "limit". For now we limit only if order is not defined.
@@ -27,7 +28,7 @@ internal sealed partial class SelectPlanner
             {
                 var fetchCount = Misc_CreateDelegate(querySpecificationNode.FetchNode.CountNode)
                     .Invoke().AsInteger;
-                foreach (var queryContext in commandContext.InputQueryContextList)
+                foreach (var queryContext in context.InputQueryContextList)
                 {
                     queryContext.QueryInfo.Limit = (queryContext.QueryInfo.Limit ?? 0) + fetchCount;
                 }
@@ -36,7 +37,7 @@ internal sealed partial class SelectPlanner
             {
                 var offsetCount = Misc_CreateDelegate(querySpecificationNode.OffsetNode.CountNode)
                     .Invoke().AsInteger;
-                foreach (var queryContext in commandContext.InputQueryContextList)
+                foreach (var queryContext in context.InputQueryContextList)
                 {
                     queryContext.QueryInfo.Limit = (queryContext.QueryInfo.Limit ?? 0) + offsetCount;
                 }
@@ -83,7 +84,7 @@ internal sealed partial class SelectPlanner
                 return false;
             }
             var valueFunc = makeDelegateVisitor.RunAndReturn(expressionNode!);
-            rowsInputContext.QueryInfo.AddCondition(column, binaryOperationExpressionNode.Operation, valueFunc);
+            commandContext.Conditions.AddCondition(column, binaryOperationExpressionNode.Operation, valueFunc);
             return true;
         }
 
@@ -108,8 +109,8 @@ internal sealed partial class SelectPlanner
             }
             var leftValueFunc = makeDelegateVisitor.RunAndReturn(betweenExpressionNode.Left);
             var rightValueFunc = makeDelegateVisitor.RunAndReturn(betweenExpressionNode.Right);
-            rowsInputContext.QueryInfo.AddCondition(column, VariantValue.Operation.GreaterOrEquals, leftValueFunc);
-            rowsInputContext.QueryInfo.AddCondition(column, VariantValue.Operation.LessOrEquals, rightValueFunc);
+            commandContext.Conditions.AddCondition(column, VariantValue.Operation.GreaterOrEquals, leftValueFunc);
+            commandContext.Conditions.AddCondition(column, VariantValue.Operation.LessOrEquals, rightValueFunc);
             return true;
         }
 
@@ -136,7 +137,7 @@ internal sealed partial class SelectPlanner
             {
                 values.Add(makeDelegateVisitor.RunAndReturn(inExpressionValue));
             }
-            rowsInputContext.QueryInfo.AddCondition(column, VariantValue.Operation.In, values.ToArray());
+            commandContext.Conditions.AddCondition(column, VariantValue.Operation.In, values.ToArray());
             return true;
         }
 
@@ -157,5 +158,31 @@ internal sealed partial class SelectPlanner
             }
         };
         callbackVisitor.Run(searchNode);
+    }
+
+    /// <summary>
+    /// Validate key columns and set values.
+    /// </summary>
+    private void QueryContext_SetKeyColumnsValues(SelectCommandContext context)
+    {
+        var rowsInputs = context.Inputs.Select(i => i.RowsInput).OfType<IRowsInputKeys>();
+
+        foreach (var input in rowsInputs)
+        {
+            foreach (var keyColumn in input.GetKeyColumns())
+            {
+                var condition = context.Conditions.FirstOrDefault(c =>
+                    Column.NameEquals(c.Column, keyColumn.ColumnName)
+                    && keyColumn.Operations.Contains(c.Operation));
+                if (condition == null && keyColumn.IsRequired)
+                {
+                    throw new QueryContextMissedCondition(keyColumn.ColumnName, keyColumn.Operations);
+                }
+                if (condition != null)
+                {
+                    input.SetKeyColumnValue(keyColumn.ColumnName, condition.ValueFunc.Invoke());
+                }
+            }
+        }
     }
 }