+++ src/QueryCat.Backend/Commands/Select/SelectPlanner.cs
@@ -75,7 +75,8 @@ internal sealed partial class SelectPlanner
         Pipeline_ApplyOffsetFetch(context, node.OffsetNode, node.FetchNode);
 
         // Fill query context.
-        QueryContext_FillQueryContextConditions(node, context);
+        QueryContext_FillQueryContextConditions(context, node);
+        QueryContext_SetKeyColumnsValues(context);
 
         // INTO.
         Pipeline_CreateOutput(context, node);