+++ src/QueryCat.Backend/Formatters/DsvFormatter.cs
@@ -3,7 +3,6 @@ using QueryCat.Backend.Abstractions;
 using QueryCat.Backend.Functions;
 using QueryCat.Backend.Storage;
 using QueryCat.Backend.Types;
-using QueryCat.Backend.Utils;
 
 namespace QueryCat.Backend.Formatters;