+++ src/QueryCat.Backend/Storage/QueryContextInputInfo.cs
@@ -1,6 +1,4 @@
 using QueryCat.Backend.Abstractions;
-using QueryCat.Backend.Relational;
-using QueryCat.Backend.Types;
 
 namespace QueryCat.Backend.Storage;